+++ src/components/Choice/index.ts
@@ -1,4 +1,4 @@
 import Choice from './Choice';
 
-export {Props, errorID, helpTextID} from './Choice';
+export {Props, helpTextID} from './Choice';
 export default Choice;